fix: fixed bugs and other minor improvements
