+++ packages/server/src/middleware/auth.js
@@ -69,16 +69,7 @@ export const authenticateToken = asyncHandler(async (req, res, next) => {
   }
 });
 
-// Check if user is super admin
-export const requireSuperAdmin = (req, res, next) => {
-  if (req.user.global_role !== 'SUPER_ADMIN') {
-    return res.status(403).json({
-      success: false,
-      message: 'Super admin access required'
-    });
-  }
-  next();
-};
+
 
 // Check if user has access to company (any role)
 export const requireCompanyAccess = asyncHandler(async (req, res, next) => {
